test: test 적절한 클래스로 이동
