+++ src/test/java/domain/TicketTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.Assertions.*;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -56,5 +57,42 @@ class TicketTest {
 		WinningNumber winningNumber = new WinningNumber(winningNumbers, bonusBall);
 
 		assertThat(ticket.getRank(winningNumber)).isEqualTo(Rank.THIRD_GRADE);
+	}
+
+	@DisplayName("Ticket 불변 객체 테스트")
+	@Test
+	void TicketImmutable() {
+		List<Integer> numbers = new ArrayList<>();
+		numbers.add(1);
+		numbers.add(2);
+		numbers.add(3);
+		numbers.add(4);
+		numbers.add(5);
+		numbers.add(6);
+
+		Ticket ticket = new Ticket(numbers);
+
+		numbers.add(7);
+		numbers.add(8);
+
+		assertThat(numbers.size()).isEqualTo(ticket.getBalls().size() + 2);
+	}
+
+	@DisplayName("Ticket getter()로 꺼냈을 때 불변이 지켜지는지 테스트")
+	@Test
+	void ticketGetterImmutable() {
+		List<Integer> numbers = new ArrayList<>();
+		numbers.add(1);
+		numbers.add(2);
+		numbers.add(3);
+		numbers.add(4);
+		numbers.add(5);
+		numbers.add(6);
+
+		Ticket ticket = new Ticket(numbers);
+		List<Ball> balls = ticket.getBalls();
+
+		assertThatThrownBy(() -> balls.add(new Ball(7)))
+			.isInstanceOf(UnsupportedOperationException.class);
 	}
 }
